Removed many warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@11004 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: a6ecc44528dd5235e4bfc0168f9cb09a59440df1

+++ scheduling/scheduler.slascheduler/src/de/uniol/inf/is/odysseus/scheduler/slascheduler/conformance/LatencyAverageConformance.java
@@ -100,7 +100,7 @@ public class LatencyAverageConformance<T> extends AbstractSLaConformance<T> {
 		IMetaAttribute metadata = metaAttributeContainer.getMetadata();
 		if (metadata instanceof ILatency) {
 			ILatency latency = (ILatency) metadata;
-			this.aggregate.addAggValue(this.nanoToMilli((double) latency
+			this.aggregate.addAggValue(this.nanoToMilli(latency
 					.getLatency()));
 		} else {
 			throw new RuntimeException("Latency missing");
